+++ src/ska_dlm_client/configdb_watcher/config.py
@@ -48,11 +48,13 @@ class WatcherArgs(CmdLineParameters):
             "--queue-connection-string",
             type=str,
             required=False,
-            help="RabbitMQ connection url",
+            default="amqp://guest:guest@rabbitmq/", # Docker value
+            help="RabbitMQ connection URL.",
         )
         self.parser.add_argument(
             "--queue-exchange-name",
             type=str,
             required=False,
+            default="dlm.outbox",
             help="RabbitMQ exchange.",
         )