+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 
 ### Added
 
+* Added configurable RabbitMQ connection string and exchange name to the Helm chart for the ConfigDB watcher.
 * ConfigDB Watcher can receive messages from the server via RabbitMQ.
 * Optional `directory_watcher.tolerations` and `directory_watcher.affinity` to `values.yaml`.
 * Test file pb-test-20260126-24294.tar.bz2 to Git LFS.